+++ src/types/index.js
@@ -1,0 +1,13 @@
+// @flow
+
+export * from './rule-config';
+export * from './fix-set-config';
+
+// import type { FixSetConfig } from './fix-set-config';
+// import { FixSetConfigSchema } from './fix-set-config';
+//
+// import type { RuleConfig } from './rule-config';
+// import { RuleConfigSchema } from './rule-config';
+//
+// export type { FixSetConfig, RuleConfig };
+// export { FixSetConfigSchema, RuleConfigSchema };