+++ test/rule.js
@@ -20,69 +20,53 @@ const testCases = {
 
   prefixSuffixes: {
     description: 'Prefix and suffixes combined',
-    config:      {
-      prefixes:      ['a', 'x', 'zul'], suffixes:      ['=', '?'], replacePrefix: true, replaceSuffix: true,
-    },
-    found: {
-      'A=':   'A', 'bdf=': 'bdf', 'a=':   '', xs:     's', 'xjd=': 'jd', 'ss?':  'ss', 'xd?':  'd', 'x?':   '', zulb:   'b',
-    },
-    notFound: { 'df=w': 'df=w', Adh: 'Adh' },
+    config:      { prefixes: ['a', 'x', 'zul'], suffixes: ['=', '?'], replacePrefix: true, replaceSuffix: true },
+    found:       { 'A=': 'A', 'bdf=': 'bdf', 'a=': '', xs: 's', 'xjd=': 'jd', 'ss?': 'ss', 'xd?': 'd', 'x?': '', zulb: 'b' },
+    notFound:    { 'df=w': 'df=w', Adh: 'Adh' },
   },
 
   exceptPrefixElements: {
     description: 'Elements and except prefixes combined',
-    config:      {
-      elements:       ['abc', 'def'], exceptPrefixes: 'a', replacePrefix:  true, replaceSuffix:  true,
-    },
-    found:    { abc: 'abc', def: 'def' },
-    notFound: { ab: 'b' },
+    config:      { elements: ['abc', 'def'], exceptPrefixes: 'a', replacePrefix: true, replaceSuffix: true },
+    found:       { abc: 'abc', def: 'def' },
+    notFound:    { ab: 'b' },
   },
 
   prefixAndexceptSuffixes: {
     description: 'Prefix and except suffixes combined',
-    config:      {
-      prefixes:       'a', exceptSuffixes: '!', replacePrefix:  true, replaceSuffix:  true,
-    },
-    found:    { a: '', ab: 'b' },
-    notFound: { 'a!': 'a', 'bf!': 'bf', jj: 'jj' },
+    config:      { prefixes: 'a', exceptSuffixes: '!', replacePrefix: true, replaceSuffix: true },
+    found:       { a: '', ab: 'b' },
+    notFound:    { 'a!': 'a', 'bf!': 'bf', jj: 'jj' },
   },
 
   exceptAndSuffixes: {
     description: 'Except and suffixes combined',
-    config:      {
-      prefixes:      'a', except:        'abc', replacePrefix: true, replaceSuffix: true,
-    },
-    found:    { a: '', ab: 'b' },
-    notFound: { abc: 'abc', sd: 'sd' },
+    config:      { prefixes: 'a', except: 'abc', replacePrefix: true, replaceSuffix: true },
+    found:       { a: '', ab: 'b' },
+    notFound:    { abc: 'abc', sd: 'sd' },
   },
 
   prefixSuffixesNoReplace: {
     description: 'Prefix and suffixes combined without replace',
     config:      { prefixes: ['a', 'x', 'zul'], suffixes: ['=', '?'] },
-    found:       {
-      'A=':   'A=', 'bdf=': 'bdf=', 'a=':   'a=', xs:     'xs', 'xjd=': 'xjd=', 'ss?':  'ss?', 'xd?':  'xd?', 'x?':   'x?', zulb:   'zulb',
-    },
-    notFound: { 'df=w': 'df=w', Adh: 'Adh' },
+    found:       { 'A=': 'A=', 'bdf=': 'bdf=', 'a=': 'a=', xs: 'xs', 'xjd=': 'xjd=', 'ss?': 'ss?', 'xd?': 'xd?', 'x?': 'x?', zulb: 'zulb' },
+    notFound:    { 'df=w': 'df=w', Adh: 'Adh' },
   },
 };
 
 // Execute test cases:
-Object.keys(testCases).forEach((key) => {
-  const testCase = testCases[key];
-
+Object.values(testCases).forEach((testCase) => {
   describe(testCase.description, () => {
     const rule = new Rule(testCase.config);
 
-    Object.keys(testCase.found).forEach((field, i) => {
-      const expected = testCase.found[field];
+    Object.entries(testCase.found).forEach(([field, expected], i) => {
       it(`should find name in sample index ${i}.`, (done) => {
         expect(rule.has(field)).to.equal({ found: true, name: expected });
         done();
       });
     });
 
-    Object.keys(testCase.notFound).forEach((field, i) => {
-      const expected = testCase.notFound[field];
+    Object.entries(testCase.notFound).forEach(([field, expected], i) => {
       it(`should not find name in sample index ${i}.`, (done) => {
         expect(rule.has(field)).to.equal({ found: false, name: expected });
         done();
@@ -93,9 +77,7 @@ Object.keys(testCases).forEach((key) => {
 
 describe('has method', () => {
   it('should allow optional replacement of prefix and suffix - 1.', (done) => {
-    const rule = new Rule({
-      prefixes:      'a', suffixes:      'z', replacePrefix: true, replaceSuffix: true,
-    });
+    const rule = new Rule({ prefixes: 'a', suffixes: 'z', replacePrefix: true, replaceSuffix: true });
     const result = rule.has('abbz', { replacePrefix: false, replaceSuffix: false });
 
     expect(result).to.equal({ found: true, name: 'abbz' });